# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AlignPath APIs",
        "description": "RESTful APIs for AlignPath, offering endpoints for authentication, user management, and personalized career recommendations with their corresponding quests, problems, skills and resources.",
        "version": "v0.1"
    },
    "paths": {
...
        "/api/user/preferences": {
            "post": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Store or update authenticated user's preferences",
                "operationId": "364b6338e3dcb1ac3976e5751a8af3da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}